+++ src/features/pythonApi.ts
@@ -303,6 +303,10 @@ export class PythonEnvironmentApiImpl implements PythonEnvironmentApi {
         }
         return manager.manage(context, options);
     }
+    async getPackageManager(context: PythonEnvironment): Promise<PackageManager | undefined> {
+        await waitForEnvManagerId([context.envId.managerId]);
+        return this.envManagers.getPackageManager(context);
+    }
     async refreshPackages(context: PythonEnvironment): Promise<Package[] | undefined> {
         await waitForEnvManagerId([context.envId.managerId]);
         const manager = this.envManagers.getPackageManager(context);